Plot of ROC Curve and Cost-Profit Curve
ROC curve is an important graph for model performance comparison, in difference of imbalanced data. Cost-Profit curve incorporates  ROC Curve have a built-in functions in sklearn, but cost-profit curve does not. You use cost-profit curve when different class labels have different importance.  A function for plotting cost-profit curve is presented in the code.